# contracts/openapi.yaml
openapi: 3.0.4
info:
  title: ConGoRAG API
  version: 0.1.0
  description: |
    依靠上下文提供 RAG 服务的全栈开源工具

    这份文件是唯一真相：Go 的接口和 TS 的类型都从它生成，
    两边不会对不上。改接口先改这里，再跑 make generate。
servers:
  - url: http://localhost:3210
    description: 本地开发

# ══════════════════════════════════════════════════════════════
# paths —— 有哪些 URL，每个支持什么方法
#
# operationId 用来当生成出来的函数名，要有业务含义且全局唯一。
# 状态码必须加引号：'200' 是字符串，200 会被解析成数字。
# ══════════════════════════════════════════════════════════════
paths:
...
  /api/v1/knowledge-bases:

    get:
      operationId: listKnowledgeBases
      summary: 列出全部知识库（按创建时间倒序）
      responses:
        '200':
          description: 知识库列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KnowledgeBase'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createKnowledgeBase
      summary: 新建一个知识库
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKnowledgeBaseRequest'
      responses:
        # 201 = 创建成功。创建类操作用 201 而不是 200。
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBase'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── 知识库：单个 ────────────────────────────────────────────
  /api/v1/knowledge-bases/{id}:

    # 路径参数写在路径这一级，三个方法共用：
    # 写在这里等于给下面三个方法都加上了这个参数，不用各写一遍。
    parameters:
      - name: id
        in: path
        required: true          # 路径参数必须是 true
        schema:
          type: string
          format: uuid

    get:
      operationId: getKnowledgeBase
      summary: 取一个知识库
      responses:
        '200':
          description: 找到了
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBase'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      operationId: renameKnowledgeBase
      summary: 改知识库的名字
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameKnowledgeBaseRequest'
      responses:
        # 204 = 成功但没有返回内容。
        #
        # 不用 200 + 返回改好的对象，是因为：
        # 那需要 usecase 的 Rename 之后再查一次库，多一次往返。
        # 而客户端本来就知道新名字（是它自己传上来的），不需要服务端再告诉它一遍。
        #
        # 如果你的前端希望直接拿到最新对象（比如要更新缓存里的 updatedAt），
        # 改成 200 + KnowledgeBase，handler 里 Rename 完再 Get 一次即可。
        '204':
          description: 改好了，无返回内容
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: deleteKnowledgeBase
      summary: 删除知识库（连带它的文档和分块）
      responses:
        '204':
          description: 删掉了，无返回内容
        '500':
          $ref: '#/components/responses/InternalError'

  # ── 文档：某个知识库下的列表 + 上传 ──────────────────────────
  /api/v1/knowledge-bases/{id}/documents:

    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      operationId: listDocuments
      summary: 列出这个知识库下的全部文档
      responses:
        '200':
          description: 文档列表（按创建时间倒序）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: uploadDocument
      summary: 上传一个文档（异步处理：落盘 + 入队，立即返回）
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        # 202 = 已接受，但处理还没完成——文档此刻是 queued 状态，
        # 真正的 parse/chunk/embed/index 由 worker 异步做。
        # 不是 201：201 暗示"资源已经完整创建好了"，这里只是排上了队。
        '202':
          description: 已接收，进入处理队列
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── 文档：单个 ──────────────────────────────────────────────
  /api/v1/documents/{id}:

    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      operationId: getDocument
      summary: 查一份文档的当前状态（前端拿它做处理中的轮询）
      responses:
        '200':
          description: 找到了
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: deleteDocument
      summary: 删除一份文档（连带它的分块；磁盘文件异步清理）
      responses:
        '204':
          description: 删掉了，无返回内容
        '500':
          $ref: '#/components/responses/InternalError'

  # ── 会话 ────────────────────────────────────────────────────
  /api/v1/conversations:

    post:
      operationId: createConversation
      summary: 新建一个会话
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/conversations/{id}/messages:

    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      operationId: listConversationMessages
      summary: 按顺序列出一个会话的全部消息（前端刷新页面重载历史用）
      responses:
        '200':
          description: 消息列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: sendMessage
      summary: |
        发一条消息，响应是 SSE 流（text/event-stream），不是普通 JSON。
        帧格式和事件类型见 docs/sse-protocol.md，那份文档是唯一权威来源，
        这里的 schema 只是占位（OpenAPI 对流式响应体的描述能力有限）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: SSE 流开始。之后的成功/失败都通过流内事件表达
            （token/citation/error/done），不再改变这个 200。
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/conversations/{id}/events:

    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      operationId: subscribeConversationEvents
      summary: 断线续传入口（SSE），docs/sse-protocol.md「断线续传」一节
      parameters:
        - name: after_event_id
          in: query
          required: false
          description: 只返回这个 id 之后的事件；省略或传 0 表示从头开始
          schema:
            type: integer
            format: int64
            default: 0
      responses:
        '200':
          description: SSE 流（补发历史事件后保持连接，继续推送新事件）
          content:
            text/event-stream:
              schema:
                type: string
        '500':
          $ref: '#/components/responses/InternalError'

  # ── BYOK：模型接入配置 ─────────────────────────────────────
  #
  # 一次 POST 做完引导页"保存并开始"那一刻的全部事情：
  # 存 provider（Key 加密落盘）→ 存 chat 模型 → 探测 embedding 维度 →
  # 存 embedding 模型 → ALTER 向量列为 halfvec(N) → 建 HNSW 索引。
  # 开发文档 §4.4 有完整流程说明。
  /api/v1/providers:

    get:
      operationId: listProviders
      summary: 列出已配置的模型接入（Key 不会出现在响应里）
      responses:
        '200':
          description: Provider 列表，每个带它挂着的模型
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProviderWithModels'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createProvider
      summary: 保存一个模型接入配置（引导页"保存并开始"）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProviderRequest'
      responses:
        '201':
          description: 探测成功，三行已落库，向量列已建好
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderWithModels'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '502':
          description: 上游模型服务出错（Base URL / Key / 模型名有问题，或探测请求失败）
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── M4-A：Agent ────────────────────────────────────────────
  /api/v1/tools:

    get:
      operationId: listToolCatalog
      summary: 列出可用工具目录（创建 Agent 表单的勾选列表用）
      responses:
        '200':
          description: 工具目录
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolCatalogEntry'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/agents:

    get:
      operationId: listAgents
      summary: 列出已创建的 Agent
      responses:
        '200':
          description: Agent 列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createAgent
      summary: 创建一个 Agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/agents/{id}:

    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      operationId: getAgent
      summary: 查一个 Agent 的配置
      responses:
        '200':
          description: Agent 详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/agents/{id}/runs:

    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      operationId: listAgentRuns
      summary: 列出一个 Agent 的历史执行记录
      responses:
        '200':
          description: Run 列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentRun'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: startAgentRun
      summary: |
        启动一次 Agent 执行，响应是 SSE 流（text/event-stream）——
        和 sendMessage 同一个模式,帧格式见 docs/sse-protocol.md。
        这一轮没有 run 维度的断线重订阅（M4-B 才做),这次连接就是
        唯一能实时看到过程的机会,执行记录本身（agent_runs/
        agent_run_steps）落库后可以事后通过 GET .../runs/{runId}/steps 查。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartAgentRunRequest'
      responses:
        '200':
          description: SSE 流开始
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/runs/{runId}/steps:

    parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      operationId: listRunSteps
      summary: 列出一次执行的全部步骤（执行轨迹页用）
      responses:
        '200':
          description: Step 列表，按 seq 升序
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentRunStep'
        '500':
          $ref: '#/components/responses/InternalError'

# ══════════════════════════════════════════════════════════════
# components —— 可复用的东西，被上面的 $ref 引用
# ══════════════════════════════════════════════════════════════
components:
...
  schemas:
    KnowledgeBase:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid        # 提示生成器：Go 用 uuid.UUID，TS 用 string
        name:
          type: string
          minLength: 1
          maxLength: 200      # 和 internal/knowledge/usecase.go 的 maxNameLen 保持一致
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── 会话 ──────────────────────────────────────────────────
    Conversation:
      type: object
      required: [id, title, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        knowledgeBaseId:
          type: string
          format: uuid
          nullable: true
          description: |
            可选关联的知识库——关联了，发消息时会做检索增强；
            不关联就是一次不带 RAG 的纯对话。方案文档没有定义这个字段，
            是实现时补上的决策（见 internal/conversation/model.go 的注释）。
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateConversationRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          example: 关于 ConGoRAG 的问答
        knowledgeBaseId:
          type: string
          format: uuid
          nullable: true
...
    Message:
      type: object
      required: [id, conversationId, role, content, status, sequenceNo, createdAt]
      properties:
        id:
          type: string
          format: uuid
        conversationId:
          type: string
          format: uuid
        role:
          type: string
          enum: [system, user, assistant, tool]
        content:
          type: string
        status:
          type: string
          enum: [streaming, completed, failed]
        sequenceNo:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time

    # ── 文档 ──────────────────────────────────────────────────
    Document:
      type: object
      required: [id, knowledgeBaseId, filename, status, byteSize, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        knowledgeBaseId:
          type: string
          format: uuid
        filename:
          type: string
          description: 用户上传时的原始文件名
        status:
          type: string
          enum: [queued, processing, ready, failed]
          description: |
            四态状态机（代码架构设计 §5.4）：queued → processing → ready，
            任意阶段失败进 failed，failed 只能退回 queued 重试。
        byteSize:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # 请求体和响应体是两个 schema，不复用 KnowledgeBase：
    # 创建时客户端只该传 name——id 和两个时间戳由服务端生成。
    # 复用的话，生成的类型里会多出三个客户端不该填的字段。
    CreateKnowledgeBaseRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: 我的知识库
    RenameKnowledgeBaseRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: 新的名字

    # ── BYOK ──────────────────────────────────────────────────
...
  responses:
    InvalidArgument:
      description: 参数不合法
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: 资源不存在
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalError:
      description: 服务内部错误
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'